Pilot
This is a simple Role-Playing Game (Or RPG) that I programmed, and it is the first game I ever programmed! I learned how to do this from Scott Lilly's tutorials, and by the time I got through it, it helped me in furthering my understanding of the C-Sharp Programming Language, where concepts, such as inheritance, constructors, static classes, if and for loops, and so on, all of which I previously learned in a classroom setting, were all meaningfully applied, to where I see them in actual practice. It was also here, that I was introduced to using SQL Server 2014. I also was introduced to LINQ and XML as well.